Hacked a fix to links and attachments not updating
The problem was(is) that the initial request for updated information
goes to the model before the model has fully updated the data.  This
results in old data being returned. The model IS updated, just after
the request for new data has already been handled.
The Hack was to put a (dumb) retry mechanism into the form updater.
It will retry n times( set to 1 now) to update the form for each
form update request by default. The retry is now a parameter of the
form update function, so the retry can be turned off by passing in
a 0 for the retyr parameter.